fix(pain001): updating tests and payment data
