+++ pyproject.toml
@@ -9,11 +9,11 @@ repository = "https://github.com/sebastienrousseau/pain001"
 homepage = "https://pain001.com"
 
 [tool.poetry.dependencies]
-click = "^8.1.3"
+click = "^8.1.7"
 defusedxml = "^0.7.1"
-python = "^3.9"
-rich = "^13.4.2"
-xmlschema = "^2.3.1"
+python = "^3.11"
+rich = "^13.5.2"
+xmlschema = "^2.4.0"
 
 [build-system]
 requires = ["poetry-core"]
