+++ tests/data/template.csv
@@ -1,6 +1,5 @@
-id,date,nb_of_txs,initiator_name,payment_information_id,payment_method,batch_booking,control_sum,service_level_code,requested_execution_date,debtor_name,debtor_account_IBAN,debtor_agent_BIC,charge_bearer,payment_id,payment_amount,currency,creditor_agent_BIC,creditor_name,creditor_account_IBAN,remittance_information,forwarding_agent_BIC
-1,2023-03-10T15:30:47.000Z,2,John Doe,Payment-Info-12345,TRF,TRUE,2,SEPA,2023-03-12,Acme Corp,DE75512108001245126162,BANKDEFFXXX,SLEV,PaymentID6789,150.00,EUR,SPUEDE2UXXX,Global Tech,DE68210501700024690959,Invoice-98765,FA98765
-2,2023-03-11T10:20:18.000Z,3,Jane Smith,Payment-Info-67890,TRF,TRUE,3,SEPA,2023-03-14,Brown Industries,DE44500105175407324931,BANKDEFFXXX,SLEV,PaymentID4321,300.00,EUR,SPUEDE2UXXX,Green Energy,DE89370400440532013008,Invoice-12345,FA24680
-3,2023-03-11T11:45:23.000Z,1,Michael Johnson,Payment-Info-24680,TRF,TRUE,1,SEPA,2023-03-13,Alpha Electronics,DE47500105175711000100,BANKDEFFXXX,SLEV,PaymentID1357,250.00,EUR,SPUEDE2UXXX,Beta Chemicals,DE44500105175407123457,Invoice-24680,FA13579
-4,2023-03-12T09:12:34.000Z,2,Emily Watson,Payment-Info-86420,TRF,TRUE,2,SEPA,2023-03-15,Delta Manufacturing,DE25500105176602214896,BANKDEFFXXX,SLEV,PaymentID2468,175.00,EUR,SPUEDE2UXXX,Gamma Logistics,DE53500105176502098765,Invoice-86420,FA86420
-5,2023-03-12T14:23:17.000Z,3,Robert Brown,Payment-Info-59162,TRF,TRUE,3,SEPA,2023-03-16,Smart Solutions,DE86500105174603245678,BANKDEFFXXX,SLEV,PaymentID4815,225.00,EUR,SPUEDE2UXXX,Eco Innovations,DE18500105175001478962,Invoice-59162,FA59162
+id,date,nb_of_txs,initiator_name,initiator_street_name,initiator_building_number,initiator_postal_code,initiator_town_name,initiator_country_code,payment_information_id,payment_method,batch_booking,requested_execution_date,debtor_name,debtor_street_name,debtor_building_number,debtor_postal_code,debtor_town_name,debtor_country_code,debtor_account_IBAN,debtor_agent_BIC,charge_bearer,payment_id,payment_amount,currency,payment_currency,control_sum,creditor_agent_BIC,creditor_name,creditor_street_name,creditor_building_number,creditor_postal_code,creditor_town_name,creditor_country_code,creditor_account_IBAN,purpose_code,reference_number,reference_date,service_level_code,forwarding_agent_BIC,remittance_information,charge_account_IBAN
+1,2023-03-10T15:30:47.000Z,2,John Doe,John's Street,1,12345,John's Town,DE,Payment-Info-12345,TRF,true,2023-03-12,Acme Corp,Acme Street,2,67890,Acme Town,DE,DE75512108001245126162,BANKDEFFXXX,SLEV,PaymentID6789,150,EUR,EUR,15000,SPUEDE2UXXX,Global Tech,Global Street,3,11223,Global Town,DE,DE68210501700024690959,OTHR,Invoice-98765,2023-03-09,SEPA,SPUEDE2UXXX,Invoice-12345,CHARGE-IBAN-12345
+2,2023-03-11T10:20:18.000Z,3,Jane Smith,Jane's Street,10,67890,Jane's Town,DE,Payment-Info-67890,TRF,true,2023-03-14,Brown Industries,Brown Street,20,45678,Brown Town,DE,DE44500105175407324931,BANKDEFFXXX,SHAR,PaymentID4321,300,EUR,EUR,30000,SPUEDE2UXXX,Green Energy,Green Street,30,78901,Green Town,DE,DE89370400440532013008,OTHR,Invoice-12345,2023-03-13,SEPA,SPUEDE2UXXX,Invoice-67890,CHARGE-IBAN-67890
+3,2023-03-11T11:45:23.000Z,1,Michael Johnson,Michael's Street,15,89101,Michael's Town,DE,Payment-Info-24680,TRF,true,2023-03-13,Alpha Electronics,Alpha Street,25,32165,Alpha Town,DE,DE47500105175711000100,BANKDEFFXXX,SLEV,PaymentID1357,250,EUR,EUR,25000,SPUEDE2UXXX,Beta Chemicals,Beta Street,35,65432,Beta Town,DE,DE44500105175407123457,OTHR,Invoice-24680,2023-03-11,SEPA,SPUEDE2UXXX,Invoice-24680,CHARGE-IBAN-24680
+4,2023-03-12T09:12:34.000Z,2,Emily Wilson,Emily's Street,4,34567,Emily's Town,DE,Payment-Info-36912,TRF,true,2023-03-15,Zeta Services,Zeta Street,8,98765,Zeta Town,DE,DE68120933300112345608,BANKDEFFXXX,SHAR,PaymentID8642,200,EUR,EUR,20000,SPUEDE2UXXX,Theta Solutions,Theta Street,38,54321,Theta Town,DE,DE75512108001245123456,OTHR,Invoice-36912,2023-03-14,SEPA,SPUEDE2UXXX,Invoice-36912,CHARGE-IBAN-36912
